refactor(cli): initial branch commit
