+++ crates/sage/src/ion_series.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use crate::mass::monoisotopic;
 use crate::peptide::Peptide;
 
-#[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Deserialize, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum Kind {
     A,
